# server/api/openapi-with-docs.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "DeadPigeons API",
    "description": "3rd Semester Project API documentation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/board/user/{userId}": {
      "get": {
        "tags": [
          "Board"
        ],
        "operationId": "Board_GetByUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardDtoResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/board/user/purchase": {
      "post": {
        "tags": [
          "Board"
        ],
        "operationId": "Board_Purchase",
        "requestBody": {
          "x-name": "dtos",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateBoardRequest"
                }
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardDtoResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/board/{boardId}/auto-repeat": {
      "put": {
        "tags": [
          "Board"
        ],
        "operationId": "Board_SetAutoRepeat",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAutoRepeatRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoRepeatResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/games/draw": {
      "post": {
        "tags": [
          "AdminGame"
        ],
        "operationId": "AdminGame_EnterWinningNumbers",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameDrawRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/games/draw/status": {
      "get": {
        "tags": [
          "AdminGame"
        ],
        "operationId": "AdminGame_IsWeekLocked",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "weekNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoardDtoResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "playerId": {
            "type": "string",
            "nullable": true
          },
          "gameId": {
            "type": "string",
            "nullable": true
          },
          "numbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "times": {
            "type": "integer",
            "format": "int32"
          },
          "repeatId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "autoRepeat": {
            "type": "boolean"
          },
          "isWinner": {
            "type": "boolean",
            "nullable": true
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardTransactionDto"
            }
          }
        }
      },
      "BoardTransactionDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "CreateBoardRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "userId",
          "numbers"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "minLength": 1
          },
          "numbers": {
            "type": "array",
            "maxItems": 8,
            "minItems": 5,
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "times": {
            "type": "integer",
            "format": "int32",
            "maximum": 2147483647.0,
            "minimum": 1.0
          }
        }
      },
      "AutoRepeatResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "boardId": {
            "type": "string"
          },
          "autoRepeat": {
            "type": "boolean"
          },
          "repeatId": {
            "type": "string",
            "nullable": true
          },
          "isStopped": {
            "type": "boolean"
          }
        }
      },
      "UpdateAutoRepeatRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "autoRepeat": {
            "type": "boolean"
          }
        }
      },
...
      "GameResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "weekNumber": {
            "type": "integer",
            "format": "int32"
          },
          "winningNumbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateGameDrawRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "weekNumber": {
            "type": "integer",
            "format": "int32"
          },
          "winningNumbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
...
    }
  }
}